+++ packages/firebase_auth/android/src/main/java/io/flutter/plugins/firebaseauth/FirebaseAuthPlugin.java
@@ -125,7 +125,6 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         handleSignOut(call, result, getAuth(call));
         break;
       case "getIdToken":
-      case "getIdTokenResult":
         handleGetToken(call, result, getAuth(call));
         break;
       case "reauthenticateWithCredential":
@@ -531,11 +530,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       return;
     }
 
-    // Contrary to iOS, android has a single method to handle getIdToken() and getIdTokenResult()
-    final boolean returnTokenResult = call.method.equals("getIdTokenResult");
-
     Map<String, Boolean> arguments = call.arguments();
-    boolean refresh = arguments.get("refresh");
+    final boolean refresh = arguments.get("refresh");
 
     currentUser
         .getIdToken(refresh)
@@ -543,24 +539,18 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
             new OnCompleteListener<GetTokenResult>() {
               public void onComplete(@NonNull Task<GetTokenResult> task) {
                 if (task.isSuccessful() && task.getResult() != null) {
-                  String idToken = task.getResult().getToken();
+                  final Map<String, Object> map = new HashMap<>();
+                  map.put("token", task.getResult().getToken());
+                  map.put("expirationTimestamp", task.getResult().getExpirationTimestamp());
+                  map.put("authTimestamp", task.getResult().getAuthTimestamp());
+                  map.put("issuedAtTimestamp", task.getResult().getIssuedAtTimestamp());
+                  map.put("claims", task.getResult().getClaims());
 
-                  if(returnTokenResult) {
-                    Map<String, Object> map = new HashMap<>();
-                    map.put("token", idToken);
-                    map.put("expirationTimestamp", task.getResult().getExpirationTimestamp());
-                    map.put("authTimestamp", task.getResult().getAuthTimestamp());
-                    map.put("issuedAtTimestamp", task.getResult().getIssuedAtTimestamp());
-                    map.put("claims", task.getResult().getClaims());
-
-                    if(task.getResult().getSignInProvider() != null) {
-                      map.put("signInProvider", task.getResult().getSignInProvider());
-                    }
-
-                    result.success(Collections.unmodifiableMap(map));
-                  } else {
-                    result.success(idToken);
+                  if(task.getResult().getSignInProvider() != null) {
+                    map.put("signInProvider", task.getResult().getSignInProvider());
                   }
+
+                  result.success(Collections.unmodifiableMap(map));
                 } else {
                   reportException(result, task.getException());
                 }